requirements.txt
# Frontend
streamlit==1.32.0
matplotlib==3.8.3

# API
flask==3.0.2
flask-restx==1.3.0
python-dotenv==1.0.1

# Backend
pymysql==1.1.0
sqlalchemy==2.0.28

# Testing
pytest==8.0.2
pytest-cov==4.1.0

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2 